requirements/base.txt
django
djangorestframework
environs
psycopg2
djangorestframework-simplejwt